+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+Version 1.02
+
+- Fix breaking pre dragonflight crafting
+- Fix for skill being higher than difficulty
+
 Version 1.01
 
 - Fix missing library import