+++ CHANGES.md
@@ -1,2 +1,4 @@
-- Fix breaking pre dragonflight crafting
-- Fix for skill being higher than difficulty
+- Exempt recipe tracker from item count override
+- Added ability to show result with inspiration bonus
+- Added ability to track multiple crafts of a recipe
+- Added searching for reagents in crafting window or AH when clicking on them in the tracker